refactor(docs): streams for getting data moved into shared/helpers
For use in other scripts because they all require access to similar pieces of state.
