+++ util/shared/helpers.mjs
@@ -21,6 +21,8 @@ import fs from 'fs'
 import path from 'path'
 import getPathOr from 'crocks/helpers/getPathOr.js'
 import { getSchema } from './json-schema.mjs'
+import { generatePropertyEvents, generatePropertySetters, generatePolymorphicPullEvents } from '../shared/modules.mjs'
+import { getExternalMarkdownPaths } from '../shared/json-schema.mjs'
 
 const {
   access,
@@ -190,7 +192,93 @@ const fileCollectionReducer = (truncateBefore = '') => (acc = {}, payload = '') 
   return acc
 }
 
+const hasPublicMethods = json => json.methods && json.methods.filter(m => !m.tags || !m.tags.map(t=>t.name).includes('rpc-only')).length > 0
+const alphabeticalSorter = (a, b) => a.info.title > b.info.title ? 1 : b.info.title > a.info.title ? -1 : 0
+const markdownFileReducer = fileCollectionReducer('/template/markdown/')
+const combineStreamObjects = (...xs) => h([...xs]).flatten().collect().map(xs => Object.assign({}, ...xs))
+const schemaMapper = ([_filepath, data]) => {
+  const parsed = JSON.parse(data)
+  if (parsed && parsed.$id) {
+    return  [parsed.$id, parsed]
+  }
+}
+
+const addExternalMarkdown = (paths = [], data = {}, descriptions = {}) => {
+  paths.map(path => {
+    const urn = getPathOr(null, path, data)
+    const url = urn.indexOf("file:../") == 0 ? urn.substr("file:../".length) : urn.substr("file:".length)
+    const markdownContent = descriptions[url]
+    path.pop() // last element is expected to be `$ref`
+    const field = path.pop() // relies on this position being the field name
+    const objectNode = getPathOr(null, path, data)
+    objectNode[field] = markdownContent // This mutates `data` by reference because JavaScript!
+  })
+  return data
+}
+
+const sharedTemplates = sharedTemplateFolder => recursiveFileDirectoryList(sharedTemplateFolder)
+  .flatFilter(isFile)
+  .through(loadFileContent('.md'))
+  .reduce({}, markdownFileReducer)
+
+const localTemplates = templateFolder => recursiveFileDirectoryList(templateFolder)
+  .flatFilter(isFile)
+  .through(loadFileContent('.md'))
+  .reduce({}, markdownFileReducer)
+
+const externalMarkdownDescriptions = markdownFolder => recursiveFileDirectoryList(markdownFolder)
+  .flatFilter(isFile)
+  .through(loadFileContent('.md'))
+  .reduce({}, fileCollectionReducer('/src/'))
+
+// TODO: Add error handling back to json docs.
+const sharedSchemas = sharedSchemasFolder => recursiveFileDirectoryList(sharedSchemasFolder)
+  .flatFilter(isFile)
+  .through(loadFileContent('.json'))
+  .map(schemaMapper)
+  .reduce({}, fileCollectionReducer())
+
+const localSchemas = schemasFolder => recursiveFileDirectoryList(schemasFolder)
+  .flatFilter(isFile)
+  .through(loadFileContent('.json'))
+  .map(schemaMapper)
+  .reduce({}, fileCollectionReducer())
+
+const localModules = (modulesFolder = '', markdownFolder = '') => recursiveFileDirectoryList(modulesFolder)
+  .flatFilter(isFile)
+  .through(loadFileContent('.json'))
+  .flatMap(([filepath, data]) => h.of(data)
+    .map(JSON.parse)
+    .map(generatePropertyEvents)
+    .map(generatePropertySetters)
+    .map(generatePolymorphicPullEvents)
+    .filter(hasPublicMethods)
+    .sortBy(alphabeticalSorter)
+    .map(transformedData => [filepath, transformedData])
+  )
+  .flatMap(payload => {
+    const [filepath, data] = payload
+    const paths = getExternalMarkdownPaths(data)
+    // Note that this only evaluates descriptions if there are any to replace in the module.
+    if (paths.length > 0) {
+      return externalMarkdownDescriptions(markdownFolder)
+        .map(descriptions => addExternalMarkdown(paths, data, descriptions))
+        .map(withExternalMarkdown => [filepath, withExternalMarkdown])
+    } else {
+      // Nothing to replace
+      return h.of(payload)
+    }
+  })
+  .reduce({}, fileCollectionReducer('/modules/'))
+
 export {
+  sharedTemplates,
+  localTemplates,
+  externalMarkdownDescriptions,
+  sharedSchemas,
+  localSchemas,
+  localModules,
+  combineStreamObjects,
   bufferToString,
   recursiveFileDirectoryList,
   clearDirectory,
